+++ ChangeLog.md
@@ -7,6 +7,9 @@
 - Added a "Focus by Year" section to the stats page using TF-IDF analysis on
   post prose content to highlight key terms per year.
   ([#554](https://github.com/davep/blogmore/pull/554))
+- Some more site generation performance improvements to offset
+  recently-added features.
+  ([#556](https://github.com/davep/blogmore/pull/556))
 
 ## v2.32.0